+++ src/test/java/io/beanmapper/testmodel/optional_getter/EntityResultWithMap.java
@@ -1,0 +1,11 @@
+package io.beanmapper.testmodel.optional_getter;
+
+import java.util.Map;
+
+public class EntityResultWithMap {
+
+    public String value;
+
+    public Map<String, EntityResultWithMap> children;
+
+}