+++ src/test/java/io/beanmapper/testmodel/optional_getter/MyEntityResultWithNestedOptionalField.java
@@ -1,0 +1,11 @@
+package io.beanmapper.testmodel.optional_getter;
+
+import java.util.Optional;
+
+public class MyEntityResultWithNestedOptionalField {
+
+    public String value;
+
+    public Optional<Optional<MyEntityResultWithNestedOptionalField>> child;
+
+}